requirements.txt
streamlit
langchain
langchain-community
langchain-openai
faiss-cpu
tiktoken
python-dateutil
